FP153 [fix] 게시글별 댓글 조회시 좋아요 눌렀는지 여부 표시

+++ src/main/java/com/develonity/board/dto/BoardResponse.java
@@ -18,14 +18,14 @@ public class BoardResponse {
   private final String nickname;
   private final String title;
   private final String content;
-  private final int boardLike;
+  private final long boardLike;
   private final LocalDateTime createdAt;
   private final LocalDateTime lastModifiedAt;
   private final List<String> imagePaths;
   private final Boolean hasLike; //로그인한 유저가 좋아요 눌렀는지 여부
 
 
-  public BoardResponse(CommunityBoard communityBoard, String nickname, int boardLike,
+  public BoardResponse(CommunityBoard communityBoard, String nickname, long boardLike,
       Boolean hasLike, List<String> imagePaths) {
     this.id = communityBoard.getId();
     this.nickname = nickname;
